+++ cmake/libstored.cmake
@@ -15,13 +15,37 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 include(CheckIncludeFileCXX)
+include(CMakeParseArguments)
 
 get_filename_component(libstored_dir_ "${CMAKE_CURRENT_LIST_DIR}" DIRECTORY)
 set(libstored_dir "${libstored_dir_}" CACHE INTERNAL "")
 
 # Create the libstored library based on the generated files.
+# Old interface: libstored_lib(libprefix libpath store1 store2 ...)
+# New interface: libstored_lib(TARGET lib DESTINATION libpath STORES store1 store1 ... [ZTH] [ZMQ|NO_ZMQ])
 function(libstored_lib libprefix libpath)
-	add_library(${libprefix}libstored STATIC
+	if("${libprefix}" STREQUAL "TARGET")
+		cmake_parse_arguments(LIBSTORED_LIB
+			"ZTH;ZMQ;NO_ZMQ"
+			"TARGET;DESTINATION"
+			"STORES"
+			${ARGV}
+		)
+	else()
+		message(DEPRECATION "Use keyword-based libstored_lib() instead.")
+		set(LIBSTORED_LIB_TARGET ${libprefix}libstored)
+		set(LIBSTORED_LIB_DESTINATION ${libpath})
+		set(LIBSTORED_LIB_STORES ${ARGN})
+	endif()
+
+	# By default use ZMQ.
+	set(LIBSTORED_LIB_ZMQ TRUE)
+
+	if(LIBSTORED_LIB_NO_ZMQ)
+		set(LIBSTORED_LIB_ZMQ FALSE)
+	endif()
+
+	add_library(${LIBSTORED_LIB_TARGET} STATIC
 		${libstored_dir}/include/stored
 		${libstored_dir}/include/stored.h
 		${libstored_dir}/include/stored_config.h
@@ -49,74 +73,74 @@ function(libstored_lib libprefix libpath)
 		${libstored_dir}/src/zmq.cpp
 	)
 
-	foreach(m IN LISTS ARGN)
-		target_sources(${libprefix}libstored PRIVATE
-			${libpath}/include/${m}.h
-			${libpath}/src/${m}.cpp)
+	foreach(m IN ITEMS ${LIBSTORED_LIB_STORES})
+		target_sources(${LIBSTORED_LIB_TARGET} PRIVATE
+			${LIBSTORED_LIB_DESTINATION}/include/${m}.h
+			${LIBSTORED_LIB_DESTINATION}/src/${m}.cpp)
 
-		set_property(TARGET ${libprefix}libstored APPEND PROPERTY PUBLIC_HEADER ${libpath}/include/${m}.h)
-		install(DIRECTORY ${libpath}/doc/ DESTINATION share/libstored)
+		set_property(TARGET ${LIBSTORED_LIB_TARGET} APPEND PROPERTY PUBLIC_HEADER ${LIBSTORED_LIB_DESTINATION}/include/${m}.h)
+		install(DIRECTORY ${LIBSTORED_LIB_DESTINATION}/doc/ DESTINATION share/libstored)
 	endforeach()
 
-	target_include_directories(${libprefix}libstored PUBLIC
+	target_include_directories(${LIBSTORED_LIB_TARGET} PUBLIC
 		$<BUILD_INTERFACE:${LIBSTORED_PREPEND_INCLUDE_DIRECTORIES}>
 		$<BUILD_INTERFACE:${libstored_dir}/include>
-		$<BUILD_INTERFACE:${libpath}/include>
+		$<BUILD_INTERFACE:${LIBSTORED_LIB_DESTINATION}/include>
 		$<INSTALL_INTERFACE:include>
 	)
 
-	string(REGEX REPLACE "^(.*)-$" "stored-\\1" libname ${libprefix})
-	set_target_properties(${libprefix}libstored PROPERTIES OUTPUT_NAME ${libname})
-	target_compile_definitions(${libprefix}libstored PUBLIC -DSTORED_NAME=${libname})
+	string(REGEX REPLACE "^(.*)-libstored$" "stored-\\1" libname ${LIBSTORED_LIB_TARGET})
+	set_target_properties(${LIBSTORED_LIB_TARGET} PROPERTIES OUTPUT_NAME ${libname})
+	target_compile_definitions(${LIBSTORED_LIB_TARGET} PUBLIC -DSTORED_NAME=${libname})
 
 	if(CMAKE_BUILD_TYPE STREQUAL "Debug")
-		target_compile_definitions(${libprefix}libstored PUBLIC -D_DEBUG=1)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PUBLIC -D_DEBUG=1)
 	else()
-		target_compile_definitions(${libprefix}libstored PUBLIC -DNDEBUG=1)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PUBLIC -DNDEBUG=1)
 	endif()
 
 	if(MSVC)
-		target_compile_options(${libprefix}libstored PRIVATE /Wall /WX)
+		target_compile_options(${LIBSTORED_LIB_TARGET} PRIVATE /Wall /WX)
 	else()
-		target_compile_options(${libprefix}libstored PRIVATE -Wall -Wextra -Werror -Wdouble-promotion -Wformat=2 -Wundef -Wconversion -ffunction-sections -fdata-sections)
+		target_compile_options(${LIBSTORED_LIB_TARGET} PRIVATE -Wall -Wextra -Werror -Wdouble-promotion -Wformat=2 -Wundef -Wconversion -ffunction-sections -fdata-sections)
 	endif()
 	if(CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-		target_compile_options(${libprefix}libstored PRIVATE -Wno-defaulted-function-deleted)
+		target_compile_options(${LIBSTORED_LIB_TARGET} PRIVATE -Wno-defaulted-function-deleted)
 	endif()
 
 	if(LIBSTORED_DRAFT_API)
-		target_compile_definitions(${libprefix}libstored PUBLIC -DSTORED_DRAFT_API=1)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PUBLIC -DSTORED_DRAFT_API=1)
 	endif()
 
 	CHECK_INCLUDE_FILE_CXX("valgrind/memcheck.h" LIBSTORED_HAVE_VALGRIND)
 	if(LIBSTORED_HAVE_VALGRIND)
-		target_compile_definitions(${libprefix}libstored PUBLIC -DSTORED_HAVE_VALGRIND=1)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PUBLIC -DSTORED_HAVE_VALGRIND=1)
 	endif()
 
-	if(TARGET libzth)
-		message(STATUS "Enable Zth integration for ${libprefix}libstored")
-		target_compile_definitions(${libprefix}libstored PUBLIC -DSTORED_HAVE_ZTH=1)
-		target_link_libraries(${libprefix}libstored PUBLIC libzth)
+	if(LIBSTORED_HAVE_ZTH AND LIBSTORED_LIB_ZTH)
+		message(STATUS "Enable Zth integration for ${LIBSTORED_LIB_TARGET}")
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PUBLIC -DSTORED_HAVE_ZTH=1)
+		target_link_libraries(${LIBSTORED_LIB_TARGET} PUBLIC libzth)
 	endif()
 
-	if(LIBSTORED_HAVE_LIBZMQ)
-		target_compile_definitions(${libprefix}libstored PUBLIC -DSTORED_HAVE_ZMQ=1)
-		target_link_libraries(${libprefix}libstored PUBLIC libzmq)
+	if(LIBSTORED_HAVE_LIBZMQ AND LIBSTORED_LIB_ZMQ)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PUBLIC -DSTORED_HAVE_ZMQ=1)
+		target_link_libraries(${LIBSTORED_LIB_TARGET} PUBLIC libzmq)
 	endif()
 
 	if(WIN32)
-		target_link_libraries(${libprefix}libstored INTERFACE ws2_32)
+		target_link_libraries(${LIBSTORED_LIB_TARGET} INTERFACE ws2_32)
 	endif()
 
 	if(LIBSTORED_HAVE_HEATSHRINK)
-		target_compile_definitions(${libprefix}libstored PUBLIC -DSTORED_HAVE_HEATSHRINK=1)
-		target_link_libraries(${libprefix}libstored PUBLIC heatshrink)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PUBLIC -DSTORED_HAVE_HEATSHRINK=1)
+		target_link_libraries(${LIBSTORED_LIB_TARGET} PUBLIC heatshrink)
 	endif()
 
 	if(${CMAKE_VERSION} VERSION_GREATER "3.6.0")
 		find_program(CLANG_TIDY_EXE NAMES "clang-tidy" DOC "Path to clang-tidy executable")
 		if(CLANG_TIDY_EXE AND LIBSTORED_CLANG_TIDY)
-			message(STATUS "Enabled clang-tidy for ${libprefix}libstored")
+			message(STATUS "Enabled clang-tidy for ${LIBSTORED_LIB_TARGET}")
 
 			string(CONCAT CLANG_TIDY_CHECKS "-checks="
 				"bugprone-*,"
@@ -172,52 +196,52 @@ function(libstored_lib libprefix libpath)
 			set(DO_CLANG_TIDY "${CLANG_TIDY_EXE}" "${CLANG_TIDY_CHECKS}"
 				"--extra-arg=-I${libstored_dir}/include"
 				"--extra-arg=-I${CMAKE_BINARY_DIR}/include"
-				"--extra-arg=-I${libpath}/include"
+				"--extra-arg=-I${LIBSTORED_LIB_DESTINATION}/include"
 				"--header-filter=.*include/libstored.*"
 				"--warnings-as-errors=*"
 				"--extra-arg=-Wno-unknown-warning-option"
 			)
 
-			set_target_properties(${libprefix}libstored
+			set_target_properties(${LIBSTORED_LIB_TARGET}
 				PROPERTIES CXX_CLANG_TIDY "${DO_CLANG_TIDY}" )
 		else()
-			set_target_properties(${libprefix}libstored
+			set_target_properties(${LIBSTORED_LIB_TARGET}
 				PROPERTIES CXX_CLANG_TIDY "")
 		endif()
 	endif()
 
 	if(LIBSTORED_ENABLE_ASAN)
-		target_compile_options(${libprefix}libstored PRIVATE -fsanitize=address -fno-omit-frame-pointer)
-		target_compile_definitions(${libprefix}libstored PRIVATE -DSTORED_ENABLE_ASAN=1)
+		target_compile_options(${LIBSTORED_LIB_TARGET} PRIVATE -fsanitize=address -fno-omit-frame-pointer)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PRIVATE -DSTORED_ENABLE_ASAN=1)
 		if(${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0")
-			target_link_options(${libprefix}libstored INTERFACE -fsanitize=address)
+			target_link_options(${LIBSTORED_LIB_TARGET} INTERFACE -fsanitize=address)
 		else()
-			target_link_libraries(${libprefix}libstored INTERFACE "-fsanitize=address")
+			target_link_libraries(${LIBSTORED_LIB_TARGET} INTERFACE "-fsanitize=address")
 		endif()
 	endif()
 
 	if(LIBSTORED_ENABLE_LSAN)
-		target_compile_options(${libprefix}libstored PRIVATE -fsanitize=leak -fno-omit-frame-pointer)
-		target_compile_definitions(${libprefix}libstored PRIVATE -DSTORED_ENABLE_LSAN=1)
+		target_compile_options(${LIBSTORED_LIB_TARGET} PRIVATE -fsanitize=leak -fno-omit-frame-pointer)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PRIVATE -DSTORED_ENABLE_LSAN=1)
 		if(${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0")
-			target_link_options(${libprefix}libstored INTERFACE -fsanitize=leak)
+			target_link_options(${LIBSTORED_LIB_TARGET} INTERFACE -fsanitize=leak)
 		else()
-			target_link_libraries(${libprefix}libstored INTERFACE "-fsanitize=leak")
+			target_link_libraries(${LIBSTORED_LIB_TARGET} INTERFACE "-fsanitize=leak")
 		endif()
 	endif()
 
 	if(LIBSTORED_ENABLE_UBSAN)
-		target_compile_options(${libprefix}libstored PRIVATE -fsanitize=undefined -fno-omit-frame-pointer)
-		target_compile_definitions(${libprefix}libstored PRIVATE -DSTORED_ENABLE_UBSAN=1)
+		target_compile_options(${LIBSTORED_LIB_TARGET} PRIVATE -fsanitize=undefined -fno-omit-frame-pointer)
+		target_compile_definitions(${LIBSTORED_LIB_TARGET} PRIVATE -DSTORED_ENABLE_UBSAN=1)
 		if(${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0")
-			target_link_options(${libprefix}libstored INTERFACE -fsanitize=undefined)
+			target_link_options(${LIBSTORED_LIB_TARGET} INTERFACE -fsanitize=undefined)
 		else()
-			target_link_libraries(${libprefix}libstored INTERFACE "-fsanitize=undefined")
+			target_link_libraries(${LIBSTORED_LIB_TARGET} INTERFACE "-fsanitize=undefined")
 		endif()
 	endif()
 
 	if(LIBSTORED_INSTALL_STORE_LIBS)
-		install(TARGETS ${libprefix}libstored EXPORT libstored
+		install(TARGETS ${LIBSTORED_LIB_TARGET} EXPORT libstored
 			ARCHIVE DESTINATION lib
 			PUBLIC_HEADER DESTINATION include
 		)
@@ -248,23 +272,53 @@ function(libstored_copy_dlls target)
 endfunction()
 
 # Generate the store files and invoke libstored_lib to create the library for cmake.
+# Old interface: libstored_generate(target store1 store2 ...)
+# New interface: libstored_generate(TARGET target [DESTINATION path] STORES store1 store2 [ZTH] [ZMQ|NO_ZMQ])
 function(libstored_generate target) # add all other models as varargs
+	if("${target}" STREQUAL "TARGET")
+		cmake_parse_arguments(LIBSTORED_GENERATE
+			"ZTH;ZMQ;NO_ZMQ"
+			"TARGET;DESTINATION"
+			"STORES"
+			${ARGV}
+		)
+	else()
+		message(DEPRECATION "Use keyword-based libstored_generate() instead.")
+		set(LIBSTORED_GENERATE_TARGET ${target})
+		set(LIBSTORED_GENERATE_STORES ${ARGN})
+	endif()
+
+	set(LIBSTORED_GENERATE_FLAGS)
+	if(LIBSTORED_GENERATE_ZTH)
+		set(LIBSTORED_GENERATE_FLAGS ${LIBSTORED_GENERATE_FLAGS} ZTH)
+	endif()
+	if(LIBSTORED_GENERATE_ZMQ)
+		set(LIBSTORED_GENERATE_FLAGS ${LIBSTORED_GENERATE_FLAGS} ZMQ)
+	endif()
+	if(LIBSTORED_GENERATE_NO_ZMQ)
+		set(LIBSTORED_GENERATE_FLAGS ${LIBSTORED_GENERATE_FLAGS} NO_ZMQ)
+	endif()
+
+	if("${LIBSTORED_GENERATE_DESTINATION}" STREQUAL "")
+		set(LIBSTORED_GENERATE_DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/libstored)
+	endif()
+
 	set(model_bases "")
 	set(generated_files "")
-	foreach(model IN ITEMS ${ARGN})
+	foreach(model IN ITEMS ${LIBSTORED_GENERATE_STORES})
 		list(APPEND models ${CMAKE_CURRENT_SOURCE_DIR}/${model})
 		get_filename_component(model_base ${model} NAME_WE)
 		list(APPEND model_bases ${model_base})
-		list(APPEND generated_files ${CMAKE_CURRENT_SOURCE_DIR}/libstored/include/${model_base}.h)
-		list(APPEND generated_files ${CMAKE_CURRENT_SOURCE_DIR}/libstored/src/${model_base}.cpp)
-		list(APPEND generated_files ${CMAKE_CURRENT_SOURCE_DIR}/libstored/doc/${model_base}.rtf)
-		list(APPEND generated_files ${CMAKE_CURRENT_SOURCE_DIR}/libstored/doc/${model_base}.csv)
-		list(APPEND generated_files ${CMAKE_CURRENT_SOURCE_DIR}/libstored/rtl/${model_base}.vhd)
-		list(APPEND generated_files ${CMAKE_CURRENT_SOURCE_DIR}/libstored/rtl/${model_base}_pkg.vhd)
+		list(APPEND generated_files ${LIBSTORED_GENERATE_DESTINATION}/include/${model_base}.h)
+		list(APPEND generated_files ${LIBSTORED_GENERATE_DESTINATION}/src/${model_base}.cpp)
+		list(APPEND generated_files ${LIBSTORED_GENERATE_DESTINATION}/doc/${model_base}.rtf)
+		list(APPEND generated_files ${LIBSTORED_GENERATE_DESTINATION}/doc/${model_base}.csv)
+		list(APPEND generated_files ${LIBSTORED_GENERATE_DESTINATION}/rtl/${model_base}.vhd)
+		list(APPEND generated_files ${LIBSTORED_GENERATE_DESTINATION}/rtl/${model_base}_pkg.vhd)
 	endforeach()
 
 	add_custom_command(
-		OUTPUT ${target}-libstored.timestamp ${generated_files}
+		OUTPUT ${LIBSTORED_GENERATE_TARGET}-libstored.timestamp ${generated_files}
 		DEPENDS ${LIBSTORED_SOURCE_DIR}/include/libstored/store.h.tmpl
 		DEPENDS ${LIBSTORED_SOURCE_DIR}/src/store.cpp.tmpl
 		DEPENDS ${LIBSTORED_SOURCE_DIR}/doc/store.rtf.tmpl
@@ -276,52 +330,48 @@ function(libstored_generate target) # add all other models as varargs
 		DEPENDS ${LIBSTORED_SOURCE_DIR}/generator/dsl/grammar.tx
 		DEPENDS ${LIBSTORED_SOURCE_DIR}/generator/dsl/types.py
 		DEPENDS ${models}
-		COMMAND ${PYTHON_EXECUTABLE} ${LIBSTORED_SOURCE_DIR}/generator/generate.py -p ${target}- ${models} ${CMAKE_CURRENT_SOURCE_DIR}/libstored
-		COMMAND ${CMAKE_COMMAND} -E touch ${target}-libstored.timestamp
-		COMMENT "Generate store from ${ARGN}"
+		COMMAND ${PYTHON_EXECUTABLE} ${LIBSTORED_SOURCE_DIR}/generator/generate.py -p ${LIBSTORED_GENERATE_TARGET}- ${models} ${LIBSTORED_GENERATE_DESTINATION}
+		COMMAND ${CMAKE_COMMAND} -E touch ${LIBSTORED_GENERATE_TARGET}-libstored.timestamp
+		COMMENT "Generate store from ${LIBSTORED_GENERATE_STORES}"
 		VERBATIM
 	)
-	add_custom_target(${target}-libstored-generate
-		DEPENDS ${target}-libstored.timestamp
+	add_custom_target(${LIBSTORED_GENERATE_TARGET}-libstored-generate
+		DEPENDS ${LIBSTORED_GENERATE_TARGET}-libstored.timestamp
 	)
 
-	libstored_lib(${target}- ${CMAKE_CURRENT_SOURCE_DIR}/libstored ${model_bases})
+	libstored_lib(TARGET ${LIBSTORED_GENERATE_TARGET}-libstored DESTINATION ${LIBSTORED_GENERATE_DESTINATION} STORES ${model_bases} ${LIBSTORED_GENERATE_FLAGS})
 
-	add_dependencies(${target}-libstored ${target}-libstored-generate)
+	add_dependencies(${LIBSTORED_GENERATE_TARGET}-libstored ${LIBSTORED_GENERATE_TARGET}-libstored-generate)
 
-	get_target_property(target_type ${target} TYPE)
+	get_target_property(target_type ${LIBSTORED_GENERATE_TARGET} TYPE)
 	if(target_type MATCHES "^(STATIC_LIBRARY|MODULE_LIBRARY|SHARED_LIBRARY|EXECUTABLE)$")
-		target_link_libraries(${target} PUBLIC ${target}-libstored)
+		target_link_libraries(${LIBSTORED_GENERATE_TARGET} PUBLIC ${LIBSTORED_GENERATE_TARGET}-libstored)
 	else()
-		add_dependencies(${target} ${target}-libstored)
+		add_dependencies(${LIBSTORED_GENERATE_TARGET} ${LIBSTORED_GENERATE_TARGET}-libstored)
 	endif()
 
-	get_target_property(target_cxx_standard ${target} CXX_STANDARD)
+	get_target_property(target_cxx_standard ${LIBSTORED_GENERATE_TARGET} CXX_STANDARD)
 	if(NOT target_cxx_standard STREQUAL "target_cxx_standard-NOTFOUND")
-		set_target_properties(${target}-libstored PROPERTIES CXX_STANDARD ${target_cxx_standard})
+		set_target_properties(${LIBSTORED_GENERATE_TARGET}-libstored PROPERTIES CXX_STANDARD ${target_cxx_standard})
 	endif()
 
 	if(${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0")
 		if(CMAKE_CXX_COMPILER_ID STREQUAL "GNU" AND NOT APPLE)
 			if(target_type STREQUAL "EXECUTABLE")
-				target_link_options(${target} PUBLIC -Wl,--gc-sections)
+				target_link_options(${LIBSTORED_GENERATE_TARGET} PUBLIC -Wl,--gc-sections)
 			endif()
 		endif()
 	endif()
 
 	if(LIBSTORED_DOCUMENTATION AND TARGET doc)
-		add_dependencies(doc ${target}-libstored-generate)
+		add_dependencies(doc ${LIBSTORED_GENERATE_TARGET}-libstored-generate)
 	endif()
 
-	libstored_copy_dlls(${target})
+	libstored_copy_dlls(${LIBSTORED_GENERATE_TARGET})
 endfunction()
 
-# libzth does not support installing yet...
-if(NOT TARGET libzth)
-	configure_file(${libstored_dir}/cmake/libstored.cmake.in ${CMAKE_BINARY_DIR}/libstored.cmake)
-	install(DIRECTORY ${libstored_dir}/include/ DESTINATION include FILES_MATCHING PATTERN "*.h")
-	install(FILES ${libstored_dir}/include/stored DESTINATION include)
-	install(EXPORT libstored DESTINATION share/libstored/cmake)
-	install(FILES ${CMAKE_BINARY_DIR}/libstored.cmake DESTINATION share/cmake/libstored)
-endif()
-
+configure_file(${libstored_dir}/cmake/libstored.cmake.in ${CMAKE_BINARY_DIR}/libstored.cmake)
+install(DIRECTORY ${libstored_dir}/include/ DESTINATION include FILES_MATCHING PATTERN "*.h")
+install(FILES ${libstored_dir}/include/stored DESTINATION include)
+install(EXPORT libstored DESTINATION share/libstored/cmake)
+install(FILES ${CMAKE_BINARY_DIR}/libstored.cmake DESTINATION share/cmake/libstored)